requirements.txt
# Core dependencies for DataPy ETL Framework
pydantic>=2.0.0
PyYAML>=6.0
click>=8.0.0
psutil>=5.9.0